+++ test/src/AsyncApiValidatorFactory.js
@@ -6,7 +6,7 @@ const AsyncApiValidator = require('../../index')
 describe('factory', () => {
   it('should throw error if file not found', async () => {
     const validator = AsyncApiValidator.fromSource('something')
-    await expect(validator).rejects.toThrowError('Error: Error opening fil')
+    await expect(validator).rejects.toThrowError('Error: Error opening file')
   })
 
   it('should throw error if unable to parse file', async () => {
@@ -25,14 +25,82 @@ describe('factory', () => {
   })
 
   it('should throw error if schema is not valid with asyncapi', async () => {
-    const validator = AsyncApiValidator.fromSource('./test/schemas/invalid.yml')
-    await expect(validator).rejects.toThrowError(new Error('data.components should NOT have additional properties'))
+    try {
+      await AsyncApiValidator.fromSource('./test/schemas/invalid.yml')
+    } catch (e) {
+      expect(e.message).toBe('schema validation failed')
+      expect(e.errors).toBe('data.components should NOT have additional properties')
+    }
   })
 
   it('should throw error if key not found', async () => {
     const validator = await AsyncApiValidator.fromSource(mocks.slack)
     const validate = () => validator.validate('lightMeasured', {})
     expect(validate).toThrowError(new Error('key lightMeasured not found'))
+  })
+
+  it('should throw error - operation is not valid', async () => {
+    const validator = await AsyncApiValidator.fromSource('./test/schemas/v1.2.0/slack.yml')
+    const validate = () => validator.validate('hello', {
+      type: null
+    }, 'device-installation-events')
+    expect(validate).toThrowError(new Error('operation "null" is not valid'))
+  })
+
+  it('should throw error - msgIdentifier is required', async () => {
+    const validator = await AsyncApiValidator.fromSource('./test/schemas/v1.2.0/slack.yml')
+    const validate = () => validator.validate('hello', {
+      type: null
+    }, 'device-installation-events', 'publish')
+    expect(validate).toThrowError(new Error('"msgIdentifier" is required with channel validation'))
+  })
+
+  it('should throw error - msgIdentifier does not exist', async () => {
+    try {
+      await AsyncApiValidator.fromSource('./test/schemas/v2.0.0/deviceMessages.yaml', {msgIdentifier: 'hello'})
+    } catch (e) {
+      expect(e.message).toBe('msgIdentifier "hello" does not exist')
+    }
+  })
+
+  it('should throw error - If schema is array and payload is object', async () => {
+    const validator = await AsyncApiValidator.fromSource('./test/schemas/v1.2.0/mqtt.yaml')
+    const validate = () => validator.validate('measurements', {
+      name: 'temperature',
+      timestamp: '2019-01-21T11:04:05Z',
+      value: 36.6
+    })
+    expect(validate).toThrowError(new Error('data should be array'))
+  })
+
+  it('should not throw error - If schema is array and payload is array', async () => {
+    const validator = await AsyncApiValidator.fromSource('./test/schemas/v1.2.0/mqtt.yaml')
+    const validate = validator.validate('measurements', [{
+      name: 'temperature',
+      timestamp: '2019-01-21T11:04:05Z',
+      value: 36.6
+    }])
+    expect(validate).toStrictEqual(true)
+  })
+
+  it('should not throw error - If schema is array and payload is object with ignoreArray option true', async () => {
+    const validator = await AsyncApiValidator.fromSource('./test/schemas/v1.2.0/mqtt.yaml', {ignoreArray: true})
+    const validate = validator.validate('measurements', {
+      name: 'temperature',
+      timestamp: '2019-01-21T11:04:05Z',
+      value: 36.6
+    })
+    expect(validate).toStrictEqual(true)
+  })
+
+  it('should not throw error - If schema is array and payload is array with ignoreArray option true', async () => {
+    const validator = await AsyncApiValidator.fromSource('./test/schemas/v1.2.0/mqtt.yaml', {ignoreArray: true})
+    const validate = validator.validate('measurements', [{
+      name: 'temperature',
+      timestamp: '2019-01-21T11:04:05Z',
+      value: 36.6
+    }])
+    expect(validate).toStrictEqual(true)
   })
 
   describe('schema 1.2.0', () => {
@@ -223,6 +291,55 @@ describe('factory', () => {
             }
           }
         })
+        expect(validate).toStrictEqual(true)
+      })
+    })
+  })
+
+  describe('schema 2.0.0', () => {
+    describe('deviceMessages', () => {
+      let validator
+      beforeEach(async () => {
+        validator = await AsyncApiValidator.fromSource('./test/schemas/v2.0.0/deviceMessages.yaml', {msgIdentifier: 'name'})
+      })
+
+      it('should throw error - channel not found', () => {
+        const validate = () => validator.validate('hello', {
+          type: null
+        }, 'device-events', 'publish')
+        expect(validate).toThrowError(new Error('channel "device-events" not found'))
+      })
+
+      it('should throw error - key not found on channel with operation', () => {
+        const validate = () => validator.validate('hello', {
+          type: null
+        }, 'device-installation-events', 'publish')
+        expect(validate).toThrowError(new Error('message with key "hello" on channel "device-installation-events" and operation "publish" not found'))
+      })
+
+      it('should throw error - data should have required property', () => {
+        const validate = () => validator.validate('DeviceInstallationResponsePublished', {
+          type: 'null'
+        }, 'device-installation-events', 'publish')
+        expect(validate).toThrowError(new Error("data should have required property 'id', data should have required property 'key', data should have required property 'generated', data should have required property 'requestId'"))
+      })
+
+      it('should not throw error', () => {
+        const validate = validator.validate('DeviceInstallationResponsePublished', {
+          id: 'bd58d14f-fd3e-449c-b60c-a56548190d68',
+          key: 'string',
+          type: 'string',
+          generated: '2017-01-09T08:27:22.222Z',
+          requestId: 'string',
+          data: {
+            deviceId: 'bd58d14f-fd3e-449c-b60c-a56548190d68',
+            id: 'bd58d14f-fd3e-449c-b60c-a56548190d68',
+            status: 'success',
+            message: 'string',
+            received: '2017-01-09T08:27:22.222Z',
+            recorded: '2017-01-09T08:27:22.222Z'
+          }
+        }, 'device-installation-events', 'publish')
         expect(validate).toStrictEqual(true)
       })
     })